requirement.txt
fastapi
uvicorn
python-dotenv
requests
transformers
torch
Pillow
